refactor: moved reload related function in a new submodule
